Move ensureAdmin out into own function
